Add files via upload
Optimization 1 displays the Winneman framing station present at ICG after removing bottlenecks upstream and downstream. Furthermore, the component builder station has been moved from 42ft away to 16ft, thus minimizing the time the framing station worker has to walk to retrieve the component. To accomplish the reduction in distance, the sub-component builder station would have to be redesigned from the U shape it was previously to just a horizontal table. This is reasonable because all components built at this station will still be able to fit on the redesigned table.