Add files via upload
Optimization 1 displays the Winneman framing station present at ICG after removing bottlenecks upstream and downstream. Furthermore, the component builder station has been moved from 42ft away to 16ft, thus minimizing the time the framing station worker has to walk to retrieve the component. However, the pathing to retrieve the component has been changed to eliminate the worker from having to climb over the component before sending it into the framing station. This will add time to the retrieval, but eliminate time in the framing process and make the overall process more safe for the worker. To accomplish the reduction in distance, the sub-component builder station would have to be redesigned from the U shape it was previously to just a horizontal table. This is reasonable because all components built at this station will still be able to fit on the redesigned table.